+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 1.0.9
 
-### Rendering
+### General
+- Simple preloader #61
 
+### Rendering
 - Simple mesh sprites #41
 - Generic mesh virtual pads #64
 - Improve Clear pass #62